Restructured the rendering and input
I realised that the rendering and the input method are tied to the GUI
and so should really be contained in just one class. I've restructured
the code prior to creating the JavaFX window to take this into account.
The game now launches into the FXGUI class, which in turn houses the
game engine. The engine receives the GUI interface to make callbacks to
the GUI throughout it's run cycle. I'm still keeping the engine entirely
seperate from the GUI in that a new GUI could be made using the GUI
interface and no code would have to be changed in the game engine class
or any of the logical classes it holds.