Added JavaFX GUI
This took a bit of re-working again, mainly by removing the game loop
from the engine. This is because the style of the game is to take input,
wait for that input to run it's course, then wait for the next input. As
such it needed to be driven by the GUI. I removed the GUI interface and
the Launch class for this reason. The input class was also pointless
since the GUI handles it now.
Note - the GameEngine class still contains all of the inner workings of
the game. The GUI class is totally dependant on it but the GameEngine
can still be plugged into other GUIs.
I have added methods into the GameEngine to allow the GUI to check if an
action is possible and then to proceed.